==> notebooks/2_match_anime_to_studio.ipynb <==
# Step 2: Matching Anime from Sales Data to MyAnimeList Studios

**Metis Project 2, Andrew Zhou**

We want to find the MyAnimeList pages of the anime from our sales data. In order to avoid having to search for and match against every anime on MyAnimeList, we link each anime to its studio on MyAnimeList and grab the studio's anime data.

import pandas as pd
import numpy as np
import sys
import time

sys.path.append('..')

from utilities.scraping_utilities import \
    get_mal_studio_info, studio_name_matcher, create_studio_df

anime_sales_df = pd.read_pickle("../data/anime_sales_df_unmatched.pickle")

Match SAT studio names to MAL studio names, then change the studio names in
the dataframe accordingly.

mal_studio_info = get_mal_studio_info()
sales_studio_names = anime_sales_df["studio"].unique()
matcher = studio_name_matcher(sales_studio_names, list(mal_studio_info.keys()))

anime_sales_df["studio"] = anime_sales_df["studio"].map(lambda x: matcher[x])
anime_sales_df.to_pickle("anime_sales_df_matched.pickle")

matched_studio_names = anime_sales_df["studio"].unique()

studio_df = create_studio_df(mal_studio_info, anime_sales_df["studio"].unique())

anime_sales_df.to_pickle("../data/anime_sales_df_matched.pickle")
studio_df.to_pickle("../data/studio_df.pickle")

studio_df
