==> notebooks/3_scrape_mal_anime_info.ipynb <==
# Step 3: Scrape MyAnimeList Information

**Metis Project 2, Andrew Zhou**

Now that we've linked the anime from our sales database to MyAnimeList, we scrape all the information we want from MAL.

from bs4 import BeautifulSoup
import pandas as pd
import requests
import re
import pickle
import time
import sys

sys.path.append('..')

from utilities.scraping_utilities import \
    get_anime_link, get_anime_data, create_mal_info_df

import pandas as pd
import numpy as np

anime_sales_df = pd.read_pickle("../data/anime_sales_df_matched.pickle")
studio_df = pd.read_pickle("../data/studio_df.pickle")

## Looking at the Studio DataFrame

Let's take a look at the format of the data.

studio_df.head(5)

studio_df.iloc[20]["anime_info"]

An example of the anime info for a studio.

## Matching Anime to MAL Links

Now let's match anime in the sales data to their MAL links. We use fuzzy matching because
some names are slightly different. Some anime fail to be matched, and we drop them. A few will match erroneously and require manual cleaning, which may be done as future work.

def get_anime_link_helper(x):
    """
    Helper for DataFrame.apply's function parameter below. Preferred to
    lambda function because of clarity.
    """
    studio_anime_info = studio_df.loc[x["studio"]]["anime_info"]
    return get_anime_link(x["title"], studio_anime_info, 
                          fuzzy_match = True, ratio = 90)

anime_sales_df["link"] = anime_sales_df.apply(get_anime_link_helper, axis=1)
anime_sales_df = anime_sales_df.dropna(subset=["link"])

## Scraping MAL Anime Data

Now we finally scrape anime data from MAL, specifying the information we want to keep.

keep_cols = ["episodes", "broadcast", "genres", "duration", "rating",\
        "score", "members", "favorites"]

mal_info_df = create_mal_info_df(anime_sales_df[["title", "link"]], keep_cols)

Since `create_mal_info_df` can fail on certain pages when we're rate-limited, we include code to find the list of anime where the query failed, scrape MAL information for only those anime, and then integrate the information with our existing dataframe. If we fail again, we can run this code repeatedly until we have scraped all our desired information.

#additional_info = create_mal_info_df(anime_sales_df[mal_info_df["score"].isnull()][["title", "link"]], keep_cols)
#mal_info_df = mal_info_df.combine_first(additional_info)

anime_sales_df.to_pickle("../data/anime_sales_df_linked.pickle") 
mal_info_df.to_pickle("../data/mal_info_df.pickle")
