==> notebooks/4_final_clean.ipynb <==
# Step 4: Final Data Cleaning and Fixes

**Metis Project 2, Andrew Zhou**

We do some final cleaning passes and add some columns to the dataframe.

from bs4 import BeautifulSoup
import pandas as pd
import numpy as np

from scipy import stats

mal_info_df = pd.read_pickle("../data/mal_info_df.pickle")
anime_sales_df = pd.read_pickle("../data/anime_sales_df_linked.pickle")

Add the MAL info.

anime_sales_df = pd.concat([anime_sales_df.loc[anime_sales_df.index], mal_info_df], join="inner", axis=1)

Parse and typecast our sales data, and make a new column.

anime_sales_df["sales"] = anime_sales_df["gross_1st_rls"].map(lambda x: x[1:].replace(",", "")).astype(float)

Drop certain unknown data.

anime_sales_df = anime_sales_df[anime_sales_df["sales"] != 0]
anime_sales_df = anime_sales_df[anime_sales_df["rating"] != "None"]
anime_sales_df = anime_sales_df[anime_sales_df["duration"] != "Unknown"]
anime_sales_df = anime_sales_df.dropna(subset=["score"])

Modify and typecast some columns and make some new ones.

cols_to_int = ["duration", "episodes", "year", "members"]
anime_sales_df[cols_to_int] = anime_sales_df[cols_to_int].astype(int)
anime_sales_df["broadcast"] = anime_sales_df["broadcast"].apply(lambda x: (None, None) if pd.isnull(x) else x)
anime_sales_df[["day", "time"]] = pd.DataFrame(anime_sales_df["broadcast"].tolist(), index=anime_sales_df.index)
anime_sales_df["hour"] = anime_sales_df["time"].apply(lambda x: int(x[:2]) if x else None)
anime_sales_df["runtime"] = anime_sales_df["episodes"]*anime_sales_df["duration"]
anime_sales_df["score"] = anime_sales_df["score"].astype(float)
anime_sales_df["favorites"] = anime_sales_df["favorites"].astype(int)

Make a binary variable (0 or 1) for each genre and add these variables to the 
dataframe.

# series of series, each series has columns of genres and 1 if it's of that genre
ser_ser = anime_sales_df["genres"].map(lambda x: pd.Series(1, index=x))
# turn series of series into a df with same indices as original df
genre_df = ser_ser.apply(lambda x: x)
genre_df = genre_df.fillna(0)

anime_sales_df = pd.concat([anime_sales_df, genre_df], axis=1)


Drop certain outliers. Decided based on exploring with the data and noting that the outliers cause significant instability in training and cross-validation scores across different train-val and k-fold CV splits.

anime_sales_df = anime_sales_df[anime_sales_df["sales"] < 500]
anime_sales_df = anime_sales_df[(np.abs(stats.zscore(anime_sales_df[["score", "duration", "members", "favorites"]])) < 2).all(axis=1)]

Put sales at the end of all the columns for neatness's sake; also remove some unnecessary columns.

column_list = anime_sales_df.columns.tolist()
column_list.remove("sales")
column_list.append("sales")

column_list = list(filter(lambda col: col not in ["avg_sales", "re_rls", "total", "gross_1st_rls", "Cars", "Dementia"], column_list))

anime_sales_df = anime_sales_df[column_list]

Remove some uncommon values for sources.

sources_by_freq = anime_sales_df.groupby("source")["title"].count().sort_values(ascending=False)
sources_by_freq

common_sources = sources_by_freq.sort_values(ascending=False).head(5).index.tolist()
anime_sales_df["source"] = anime_sales_df["source"].apply(lambda x: x if x in common_sources else None)

anime_sales_df.head()

anime_sales_df.columns

anime_sales_df.to_pickle("../data/anime_sales_df_final.pickle")
